support for subtree handlers
api clients can now create a handler for an entire subtree instead of
just a singular oid. The handler still only returns one varbind, however
of a request comes through for an oid that is within the subtree root
the client handler will be called with the particular oid that was
requested. it is up to the client code to determine what varbind to
return within the subtree.

Both get and get-next now call down into the same underlying recursive
search routine.