do not report underlying bridge as interface
The logic prior to this commit considered any interface that was a part
of a bridge to be a bridged interface. While this is certainly true,
within the context of VLAN control through SNMP-QBridge this is an
inappropriate semantic understanding of what constitutes a bridge. A
bridge is made up of its interface members, and that is it.

This commit introduces a mechanism to filter for physical devices only
when reporting bridged devices, to the virtual bridge itself does not
get included.